+++ projects/web/src/components/App/index.js
@@ -1,0 +1,3 @@
+import AppContainer from './AppContainer.jsx'
+
+export default AppContainer