+++ projects/web/src/components/Todo/Alert/index.js
@@ -1,0 +1,3 @@
+import Alert from './Alert.jsx'
+
+export default Alert