+++ projects/web/src/components/Todo/ListItem/index.js
@@ -1,0 +1,3 @@
+import ListItemContainer from './ListItemContainer.jsx'
+
+export default ListItemContainer