+++ projects/web/src/components/Todo/index.js
@@ -1,0 +1,3 @@
+import Todo from './Todo.jsx'
+
+export default Todo;