+++ projects/web/src/components/TopNavigation/index.js
@@ -1,0 +1,3 @@
+import TopNavigationContainer from './TopNavigationContainer.jsx'
+
+export default TopNavigationContainer