+++ projects/web/src/config/prod.js
@@ -1,0 +1,5 @@
+module.exports = {
+    api: {
+        root: 'http://localhost:3000/demo/'
+    }
+};