+++ projects/web/src/reducers/App.js
@@ -1,0 +1,7 @@
+export default function Reducers(state = {
+    env: undefined,
+    publicPath: undefined,
+    config: undefined
+}, action) {
+    return state;
+}