+++ projects/web/webpack.server.fs.config.js
@@ -1,0 +1,1 @@
+module.exports = require('./tools/webpack.server.config')('fs');